Replace input_text by entity in get_data
